Before, running the system tests verbosely with timing, top three slowest tests running on my local workstation:
90.16s call     bigquery/tests/system.py::TestBigQuery::test_query_many_columns
60.80s call     bigquery/tests/system.py::TestBigQuery::test_load_table_from_file_w_explicit_location
34.53s call     bigquery/tests/system.py::TestBigQuery::test_list_rows_many_columns

With these changes:
30.34s call     bigquery/tests/system.py::TestBigQuery::test_query_many_columns
21.44s call     bigquery/tests/system.py::TestBigQuery::test_load_table_from_file_w_explicit_location
(removed)      bigquery/tests/system.py::TestBigQuery::test_list_rows_many_columns


Changes to test_query_many_columns:

Previously, the wide column tests generated row data for a large (10k) column table locally, sent it via a POST-based load job, and then ran a rowread test and a query test against it.  It also depended on some nondeterministic behaviors such as row ordering remaining stable.

New behavior is to leverage a DDL CTAS query that generates the desired number of rows, with the field values being generated in a deterministic fashion based on value within the first field. Testing list_rows() on the raw table vs a SELECT * isn't a meaningful distinction as list_rows is tested elsewhere, so behavior differences will come out elsewhere.


Changes to  test_load_table_from_file_w_explicit_location:

We know (and test) that jobs work when data and job location are aligned, so save some time on the test by excising the additional verification using explicit location and instead focus on verifying the failures.  We do keep a query test in the correct location, but turn it into a dry run since we can validate location mismatches in dry_run stage rather than needing to invoke a full query.