@tseaver 2018-05-10
@shollyman Thanks muchly for working to reduce the BQ system test times!